upgrade to neon 1.0 (napi-8)
- note that this drops support for node v10 and requires point-release updates for 12 & 14

+++ src/context/api.rs
@@ -921,7 +921,7 @@ pub fn measureText(mut cx: FunctionContext) -> JsResult<JsArray> {
   let width = opt_float_arg(&mut cx, 2);
   let text_metrics = this.measure_text(&text, width);
 
-  let results = JsArray::new(&mut cx, text_metrics.len() as u32);
+  let results = JsArray::new(&mut cx, text_metrics.len());
   for (i, info) in text_metrics.iter().enumerate(){
     let line = floats_to_array(&mut cx, info)?;
     results.set(&mut cx, i as u32, line)?;
