+++ src/path.rs
@@ -505,7 +505,7 @@ pub fn edges(mut cx: FunctionContext) -> JsResult<JsArray> {
 
     if let Some(edge) = from_verb(verb){
       let cmd = cx.string(edge);
-      let segment = JsArray::new(&mut cx, 1 + points.len() as u32);
+      let segment = JsArray::new(&mut cx, 1 + points.len());
       segment.set(&mut cx, 0, cmd)?;
 
       let at_point = if points.len()>1{ 1 }else{ 0 };
@@ -526,7 +526,7 @@ pub fn edges(mut cx: FunctionContext) -> JsResult<JsArray> {
     }
   }
 
-  let verbs = JsArray::new(&mut cx, edges.len() as u32);
+  let verbs = JsArray::new(&mut cx, edges.len());
   for (i, segment) in edges.iter().enumerate(){
     verbs.set(&mut cx, i as u32, *segment)?;
   }
